requirements.txt
# Discord Bot Dependencies
discord.py>=2.3.0
python-dotenv>=1.0.0

# Data Processing & Export
pandas>=2.0.0
openpyxl>=3.1.0

# Configuration
PyYAML>=6.0.1

# Database
aiosqlite>=0.19.0

# Utilities
python-dateutil>=2.8.2
